feat(chinba): 데스크톱에서 친바 만들기·시간 입력을 팀 상세 프레임에 임베드
- 팀 상세(/chinba/team/detail)에 ?view=create / ?view=event&eventId= 임베드 뷰 추가
  — 세그먼트 탭·운영 패널을 유지한 채 본문 컬럼 안에서 렌더 (데스크톱 전용)
- 모바일·딥링크는 기존 풀페이지 라우트 유지, 모바일 뷰포트로 진입/리사이즈 시
  동등한 풀페이지 URL로 replace
- TeamEventCreateView/ChinbaDetailClient에서 본문을 TeamEventCreateBody/
  ChinbaEventDetailBody로 추출해 풀페이지·임베드 양쪽에서 재사용
- DateSelector max-w 캡, ChinbaScheduleGrid/HeatmapGrid 셀 폭 상한(64px)으로
  넓은 화면에서 달력·시간 그리드가 과대해지는 문제 수정
- 공유/링크 복사는 임베드 URL이 아닌 정규 /chinba/event URL로 생성
