+++ app/(main)/chinba/_components/team/TeamDetailView.tsx
@@ -4,12 +4,15 @@ import { useState, useEffect } from 'react';
 
 import { useRouter, useSearchParams } from 'next/navigation';
 import { FiSettings } from 'react-icons/fi';
+import { LuChevronLeft } from 'react-icons/lu';
 
 import ClubSwitcher from '@/(main)/chinba/_components/team/ClubSwitcher';
 import TeamCategoriesModal from '@/(main)/chinba/_components/team/TeamCategoriesModal';
+import TeamEventCreateBody from '@/(main)/chinba/_components/team/TeamEventCreateBody';
 import TeamGroupsModal from '@/(main)/chinba/_components/team/TeamGroupsModal';
 import TeamMembersModal from '@/(main)/chinba/_components/team/TeamMembersModal';
 import TeamOpsPanel from '@/(main)/chinba/_components/team/TeamOpsPanel';
+import ChinbaEventDetailBody from '@/(main)/chinba/event/_components/ChinbaEventDetailBody';
 import CategoryFilterBar from '@/(main)/teams/_components/CategoryFilterBar';
 import GroupFilterBar from '@/(main)/teams/_components/GroupFilterBar';
 import TeamSegmentTabs, { type TeamSegment } from '@/(main)/teams/_components/TeamSegmentTabs';
@@ -25,6 +28,7 @@ import { useSmartBack } from '@/_lib/hooks/useSmartBack';
 import { useTeamDetail } from '@/_lib/hooks/useTeam';
 import { setLastTeamId } from '@/_lib/utils/chinbaSelection';
 import { canEditTeam } from '@/_lib/utils/teamPermissions';
+import { DESKTOP_MEDIA_QUERY, isDesktopViewport } from '@/_lib/utils/viewport';
 import { useAuthInitialized } from '@/providers';
 
 export default function TeamDetailView() {
@@ -35,6 +39,18 @@ export default function TeamDetailView() {
   const teamId = Number(searchParams.get('id'));
   const tabParam = searchParams.get('tab') as TeamSegment | null;
   const initialTab: TeamSegment = tabParam === 'mwoheni' || tabParam === 'jabahbwa' ? tabParam : 'mannaja';
+
+  // 데스크톱 임베드 뷰 (?view=create | ?view=event&eventId=..) — 만들기/이벤트 상세를
+  // 팀 상세 프레임(세그먼트 탭 + 운영 패널) 안에서 렌더한다. 모바일은 풀페이지 라우트 유지.
+  const viewParamRaw = searchParams.get('view');
+  const viewEventId = searchParams.get('eventId') || '';
+  const view: 'create' | 'event' | null =
+    viewParamRaw === 'create' ? 'create' : viewParamRaw === 'event' && viewEventId ? 'event' : null;
+  const viewSetId = searchParams.get('setId') ? Number(searchParams.get('setId')) : null;
+  const viewGroupId = searchParams.get('groupId') ? Number(searchParams.get('groupId')) : null;
+  const viewCategoryIdRaw = searchParams.get('categoryId');
+  const viewCategoryId =
+    viewCategoryIdRaw && Number.isFinite(Number(viewCategoryIdRaw)) ? Number(viewCategoryIdRaw) : null;
 
   const isAuthReady = useAuthInitialized();
   const { data: team, isLoading, isError } = useTeamDetail(teamId);
@@ -72,6 +88,33 @@ export default function TeamDetailView() {
     if (teamId) setLastTeamId(teamId);
   }, [teamId]);
 
+  // 임베드 뷰는 데스크톱 전용 — 모바일 뷰포트(딥링크·리사이즈)면 동등한 풀페이지 라우트로 치환
+  useEffect(() => {
+    if (!view || !teamId) return;
+    const toFullPage = () => {
+      if (view === 'create') {
+        const params = new URLSearchParams({ id: String(teamId) });
+        if (viewSetId) params.set('setId', String(viewSetId));
+        if (viewGroupId) params.set('groupId', String(viewGroupId));
+        if (viewCategoryId) params.set('categoryId', String(viewCategoryId));
+        router.replace(`/chinba/team/event-create?${params.toString()}`);
+      } else {
+        const returnTo = encodeURIComponent(`/chinba/team/detail?id=${teamId}&tab=mwoheni`);
+        router.replace(`/chinba/event?id=${viewEventId}&returnTo=${returnTo}`);
+      }
+    };
+    if (!isDesktopViewport()) {
+      toFullPage();
+      return;
+    }
+    const mql = window.matchMedia(DESKTOP_MEDIA_QUERY);
+    const onChange = (e: MediaQueryListEvent) => {
+      if (!e.matches) toFullPage();
+    };
+    mql.addEventListener('change', onChange);
+    return () => mql.removeEventListener('change', onChange);
+  }, [view, teamId, viewEventId, viewSetId, viewGroupId, viewCategoryId, router]);
+
   // Sync tab state when URL changes (e.g. browser back/forward)
   useEffect(() => {
     const tab = searchParams.get('tab') as TeamSegment | null;
@@ -102,6 +145,34 @@ export default function TeamDetailView() {
 
   const handleSettingsClick = () => {
     router.push(`/chinba/team/settings?id=${teamId}`);
+  };
+
+  // 만들기/이벤트 상세 진입 — 데스크톱은 임베드(?view=), 모바일은 기존 풀페이지 라우트
+  const goCreate = () => {
+    const params = new URLSearchParams({ id: String(teamId) });
+    if (effectiveSetId) params.set('setId', String(effectiveSetId));
+    if (selectedGroupId) params.set('groupId', String(selectedGroupId));
+    if (selectedCategoryId) params.set('categoryId', String(selectedCategoryId));
+    if (isDesktopViewport()) {
+      params.set('tab', 'mannaja');
+      params.set('view', 'create');
+      router.push(`/chinba/team/detail?${params.toString()}`);
+    } else {
+      router.push(`/chinba/team/event-create?${params.toString()}`);
+    }
+  };
+
+  const goEvent = (eventId: string) => {
+    if (isDesktopViewport()) {
+      router.push(`/chinba/team/detail?id=${teamId}&tab=mannaja&view=event&eventId=${eventId}`);
+    } else {
+      const returnTo = encodeURIComponent(`/chinba/team/detail?id=${teamId}&tab=mwoheni`);
+      router.push(`/chinba/event?id=${eventId}&returnTo=${returnTo}`);
+    }
+  };
+
+  const closeView = () => {
+    router.replace(`/chinba/team/detail?id=${teamId}&tab=mannaja`);
   };
 
   if (!teamId) {
@@ -142,7 +213,7 @@ export default function TeamDetailView() {
 
   const renderTabContent = () => {
     if (activeTab === 'mannaja') {
-      return <MannajaTab teamId={teamId} myRole={team.my_role} memberCount={team.member_count} inviteCode={team.invite_code} selectedSetId={effectiveSetId} selectedGroupId={selectedGroupId} selectedCategoryId={selectedCategoryId} terminology="club" />;
+      return <MannajaTab teamId={teamId} myRole={team.my_role} memberCount={team.member_count} inviteCode={team.invite_code} selectedSetId={effectiveSetId} selectedGroupId={selectedGroupId} selectedCategoryId={selectedCategoryId} terminology="club" onCreateEvent={goCreate} onOpenEvent={goEvent} />;
     }
     if (activeTab === 'mwoheni') {
       return <ActivityTab teamId={teamId} myRole={team.my_role} selectedSetId={effectiveSetId} selectedGroupId={selectedGroupId} selectedCategoryId={selectedCategoryId} terminology="club" />;
@@ -180,30 +251,73 @@ export default function TeamDetailView() {
             <TeamSegmentTabs activeTab={activeTab} onTabChange={handleTabChange} />
           </div>
 
-          {/* Group / Category Filter */}
-          {(groupSets.length > 0 || categories.length > 0) && (
-            <div className="shrink-0 px-4 pt-3 space-y-2">
-              {groupSets.length > 0 && (
-                <GroupFilterBar
-                  groupSets={groupSets}
-                  selectedSetId={effectiveSetId}
-                  selectedGroupId={selectedGroupId}
-                  onSetChange={setSelectedSetId}
-                  onGroupChange={setSelectedGroupId}
-                />
+          {view ? (
+            /* 데스크톱 임베드 뷰 — 만들기 폼 / 이벤트 상세를 프레임 안에서 렌더 */
+            <>
+              <div className="shrink-0 flex items-center gap-1 px-2 pt-3">
+                <button
+                  onClick={closeView}
+                  className="group rounded-full p-1.5 text-gray-600 transition-all hover:bg-gray-100 hover:text-gray-900 active:scale-95"
+                  aria-label="목록으로"
+                >
+                  <LuChevronLeft size={20} strokeWidth={2.5} className="transition-transform group-hover:-translate-x-0.5" />
+                </button>
+                <span className="text-sm font-bold text-gray-800">
+                  {view === 'create' ? '동아리 친바 만들기' : '일정 상세'}
+                </span>
+              </div>
+              <div className="flex min-h-0 flex-1 flex-col pt-3">
+                {view === 'create' ? (
+                  <TeamEventCreateBody
+                    teamId={teamId}
+                    preSetId={viewSetId}
+                    preGroupId={viewGroupId}
+                    preCategoryId={viewCategoryId}
+                    onSuccess={closeView}
+                  />
+                ) : (
+                  <ChinbaEventDetailBody
+                    key={viewEventId}
+                    eventId={viewEventId}
+                    variant="embedded"
+                    onDeleted={closeView}
+                    onCompleted={(recordQuery) =>
+                      router.replace(
+                        `/chinba/team/detail?id=${teamId}&tab=mwoheni${recordQuery ? `&${recordQuery}` : ''}`
+                      )
+                    }
+                  />
+                )}
+              </div>
+            </>
+          ) : (
+            <>
+              {/* Group / Category Filter */}
+              {(groupSets.length > 0 || categories.length > 0) && (
+                <div className="shrink-0 px-4 pt-3 space-y-2">
+                  {groupSets.length > 0 && (
+                    <GroupFilterBar
+                      groupSets={groupSets}
+                      selectedSetId={effectiveSetId}
+                      selectedGroupId={selectedGroupId}
+                      onSetChange={setSelectedSetId}
+                      onGroupChange={setSelectedGroupId}
+                    />
+                  )}
+                  <CategoryFilterBar
+                    categories={categories}
+                    selectedCategoryId={selectedCategoryId}
+                    onChange={setSelectedCategoryId}
+                  />
+                </div>
               )}
-              <CategoryFilterBar
-                categories={categories}
-                selectedCategoryId={selectedCategoryId}
-                onChange={setSelectedCategoryId}
-              />
-            </div>
-          )}
 
-          {/* Tab Content */}
-          <div className="flex-1 min-h-0 overflow-y-auto px-4 py-6">
-            {renderTabContent()}
-          </div>
+              {/* Tab Content */}
+              <div className="flex-1 min-h-0 overflow-y-auto px-4 py-6">
+                {renderTabContent()}
+              </div>
+            </>
+          )}
         </div>
 
         {/* Desktop-only operations panel (운영진 only) */}
@@ -214,7 +328,7 @@ export default function TeamDetailView() {
             onOpenMembers={() => setShowMembers(true)}
             onOpenGroups={() => setShowGroups(true)}
             onOpenCategories={() => setShowCategories(true)}
-            onCreateEvent={() => router.push(`/chinba/team/event-create?id=${teamId}`)}
+            onCreateEvent={goCreate}
             onRecordActivity={() => handleTabChange('mwoheni')}
           />
         )}
